refactor(request): extract approval outcome posting helper
